divide files for ease of reading, embed for fights, upgrades, info, improved upgrade mechanics, emotes!
