+++ ship_file.json
@@ -1,1 +1,1 @@
-[{"captain": "bdavs", "cannons": 15, "crew": 5, "armor": 5, "sails": 5, "gold": 950}, {"captain": "OwO", "cannons": 5, "crew": 5, "armor": 5, "sails": 5, "gold": 900, "position": 1}, {"captain": "MrBones1102", "cannons": 5, "crew": 10, "armor": 5, "sails": 5, "gold": 49, "position": 2}]
+[{"captain": "OwO", "cannons": 5, "crew": 5, "armor": 5, "sails": 5, "gold": 900, "position": 0}, {"captain": "MrBones1102", "cannons": 5, "crew": 10, "armor": 5, "sails": 5, "gold": 49, "position": 2}, {"captain": "bdavs", "cannons": 10, "crew": 11, "armor": 5, "sails": 5, "gold": 7605}]
